Support Cargo workspaces
Under a Cargo workspace, the 'doc' directory is found at the 'target'
directory of the workspace root, rather than at the 'target' directory
of the crate. This change implements similar logic to how Cargo finds
where to place the 'target' directory by default, so it should for both
non-workspace and workspace environments.